% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal.R
\name{compound}
\alias{compound}
\title{Compounding and Discounting}
\usage{
compound(
  A = NULL,
  r = NULL,
  n = NULL,
  interval = 1L,
  simple = T,
  continuous = F
)
}
\arguments{
\item{A}{initial amount}

\item{r}{interest rate (in percent)}

\item{n}{compounding steps, or compounding time horizon. For example, a yearly compounding would be n = 1, with
an interval of interval = 1. Wheras a bi-yearly compounding over 1 year would be n = 1, interval = 2;}

\item{interval}{the compounding interval (integer number) dividing the compounding steps n. Default 1L}

\item{simple}{Use linear (default) or exponential compounding}

\item{continuous}{should we use continuous compounding}
}
\value{
returns the compounded volume
}
\description{
compounds some initial invest A for some time steps n using the
interest rate r
}
\author{
christian bitter
}
