% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{conditional_sharpe_ratio}
\alias{conditional_sharpe_ratio}
\title{Risk}
\usage{
conditional_sharpe_ratio(
  log_returns,
  risk_free_return,
  trading_days = 250,
  alpha = 0.05
)
}
\description{
Risk
}
\author{
Christian Bitter
}
