% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal.R
\name{discount}
\alias{discount}
\title{Compounding and Discounting}
\usage{
discount(A = NULL, r = NULL, n = NULL, interval = 1L, simple = F)
}
\arguments{
\item{A}{initial amount}

\item{r}{interest rate (in percent)}

\item{n}{discounting steps}

\item{interval}{the discounting interval (integer number) dividing the discounting steps n. Default 1L}

\item{simple}{Use linear (simple) or exponential (non-simple, default) discounting}
}
\value{
returns the discounted volume
}
\description{
discounts some initial invest A for some time steps n using the
interest rate r (to derive the discount factor d)
}
\author{
christian bitter
}
