% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal.R
\name{effective_rate}
\alias{effective_rate}
\title{Compounding and Discounting}
\usage{
effective_rate(r = NULL, interval = 1L, continuous = F)
}
\arguments{
\item{r}{nominal interest rate (in percent)}

\item{interval}{interval as 1/interval when compounding is performed}

\item{continuous}{should we use continuous compounding}
}
\value{
the effective rate for the given nominal rate
}
\description{
computes the effective rate for the provided nominal rate and compouding
frequency.
}
\author{
christian bitter
}
