% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cashflow.R
\name{future_value}
\alias{future_value}
\title{Cashflows}
\usage{
future_value(
  cashflow = NULL,
  r = NULL,
  terminal_period = NULL,
  simple = T,
  verbose = T
)
}
\arguments{
\item{cashflow}{a collection of cashflows made to an idealized bank - a cashflow stream (positive and negative)}

\item{r}{interest rate (in percent)}

\item{terminal_period}{after how many future periods the cashflow should be realized. if this is zero, then
the terminal period is the last time step a cashflow transaction occurs (e.g., length(cashflow)).}

\item{simple}{Use linear (default) or exponential discounting}

\item{verbose}{should verbose logging be used}
}
\value{
returns the value of the cashflow after the terminal_period
}
\description{
computes the future value of some cashflow stream using the provided interest rate,
cashflow transaction, compounding type and termin period.
}
\author{
christian bitter
}
